// src/pages/general/UpdateData/forms/EconomicActivityForm/types.ts
interface IEconomicActivityEntry {
  economicActivity: string;
  profession: string;
  job: string;
  mainCiiuActivity: string;
  secondaryCiiuActivity: string;
  economicSector: string;
  company: string;
  contractType: string;
  admissionDate: string;
  contractExpiration: string;
  severanceRegime: string;
  workday: string;
  position: string;
  dependence: string;
  employeeCode: string;
  companyFormality: string;
  companyCountry: string;
  companyCity: string;
  companyPhone: string;
  companyAddress: string;
  companyEmail: string;
}

export type { IEconomicActivityEntry };